requirements.txt
youtube-transcript-api
pytubefix
requests
python-dotenv
google-genai
google-api-python-client
openai
pandas
tqdm

# For image processing
Pillow